feat: ReviewComponent.jsx for better modularity
